# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/firebase": {
      "post": {
        "operationId": "AuthController_firebase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/cart": {
      "get": {
        "operationId": "CartController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cart"
        ]
      },
      "post": {
        "operationId": "CartController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/{productId}": {
      "patch": {
        "operationId": "CartController_set",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "operationId": "CartController_remove",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cart"
        ]
      }
    },
...
    "/api/v1/payments/checkout": {
      "post": {
        "operationId": "PaymentsController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/confirm": {
      "post": {
        "operationId": "PaymentsController_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/fail": {
      "post": {
        "operationId": "PaymentsController_fail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/webhook/razorpay": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [
          {
            "name": "x-razorpay-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "DGkart API",
    "description": "REST API shared by the DKKart mobile app and the future DGkart.com website.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FirebaseLoginDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9910123503"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9910123503"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
...
      "AddToCartDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "object",
            "default": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "QuantityDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "quantity"
        ]
      },
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          }
        },
        "required": [
          "orderId"
        ]
      },
      "ConfirmDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "gatewayOrderId": {
            "type": "string"
          },
          "paymentId": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "gatewayOrderId",
          "paymentId",
          "signature"
        ]
      },
...
    }
  }
}